gps_system/pothole/requirements.txt
ultralytics
onnxruntime
opencv-python
numpy
pyyaml
pandas
matplotlib
tqdm
scikit-learn
